+++ src/test/java/org/scion/jpan/demo/ScmpTracerouteDemo.java
@@ -52,7 +52,7 @@ public class ScmpTracerouteDemo {
     this.localPort = localPort;
   }
 
-  private static final Network network = Network.SCION_PROTO;
+  private static final Network network = Network.PRODUCTION;
 
   public static void main(String[] args) throws IOException {
     switch (network) {